blog 1.0.0

+++ modules/core/database/migrations/2024_02_13_112720_create_verification_codes_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('code');
             $table->string('tableable_type');
             $table->foreignId('tableable_id')->nullable();
-            $table->integer('created_at');
-            $table->integer('updated_at');
+            $table->bigInteger('created_at');
+            $table->bigInteger('updated_at');
         });
     }
